% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgCheckInput.R
\docType{package}
\name{pgCheckInput-package}
\alias{pgCheckInput-package}
\alias{_PACKAGE}
\title{pgCheckInput}
\description{
Package with functionality for checking common input errors from R to Bionavigator. 
Common error ID and PamCloud links.
}
